Refactor repo organization
